editoast: remove always-on directive
This modification was done in order to bring logs for any request that comes in
thanks to the `tower_http::trace::TraceLayer` middleware. But the code has a
sneaky behavior where it's actually impossible to turn it of anymore.

For example, `RUST_LOG=info,tower_http=none` will still show `tower_http` logs.

Let's remove it.

Signed-off-by: Jean SIMARD <woshilapin@tuziwo.info>