front: use matchPathStepAndOp() in TimesStopsInput's clearPathStep()
We already have a function which matches an OP and a path step.
Re-use it instead of hand-rolling our own logic.

In the UIC case, match by positionOnPath instead of name: this
should be more robust. In the track offset case, drop the opId
check because track offsets never have an OP ID.

Signed-off-by: Simon Ser <contact@emersion.fr>