front: fix NGE path steps in updatePathStepsFrom tests
Trains created via NGE don't have a secondary_code field in their
PathSteps passed to updatePathStepsFromOperationalPoints(). This
can be checked by creating a train in NGE, setting a rolling stock
for it, and then adding a console.log() right before the
updatePathStepsFromOperationalPoints() call. The ch field is
populated and the secondary_code field doesn't exist.

This is because all PathItemLocation entries are piped into
computeBasePathSteps(), which strips secondary_code regardless of
whether the path steps are defined in terms of UIC (by OSRD) or
trigram (by NGE).

Signed-off-by: Simon Ser <contact@emersion.fr>