core: avoid recomputing unnecessary coastings, which sometimes crashed
The forward coasting would sometimes not intersect with the
base curve, which causes some weirdness and 0-length steps.

That coasting would only happen if there's no intersection with
min speed limit. In that case, it's not actually necessary
to compute it (it's supposed to give the same result as the
backwards one).

Signed-off-by: Eloi Charpentier <eloi.charpentier.42@gmail.com>