front: fix train edition setup
Since recently, when opening the view to edit a train, a pathfinding request is automatically launched. But it is launched
before the store update, which creates the bug: the pathfinding is launched with the pathsteps stored in the store (and not the ones
of the train we want to edit) and once the pathfinding request is finished, we update the store with the result of the request (and erase
the path steps of the currently selected train).

To fix this, we can directly update the store before opening the ManageTrainSchedule view.

Signed-off-by: Clara Ni <clara.ni@outlook.fr>