editoast: refactor test in a single test with cases
All the existing test follow the same pattern.
Since we do have `rstest` that allows us to express them
as cases of the same test, let's use that.

In order to shorten (and, subjectively, make it more readable and
concise?), I also implemented a `FromStr` for `TrackRange`. It's a nice
helper to express all the tests more concisely.

Signed-off-by: Jean SIMARD <woshilapin@tuziwo.info>