front: fix lazy loading for rolling stock images
We were unconditionally fetching the rolling stock image via
getDocument(). Instead, side-step getDocument() entirely, and
directly use the document URL as the image source. That way, the
loading="lazy" attribute can be used and the image is only fetched
if it's visible.

The deleted comment is wrong: authentication is required, but is
performed via a cookie. Cookies are always sent for same-origin
image requests.

Signed-off-by: Simon Ser <contact@emersion.fr>