front: use Date instead of string in OsrdConfState.startTime
Improves type safety.

Previously, it was important for OsrdConfState.startTime to be in
a specific timezone, because the string was later truncated and
directly used on an datetime-local form input. This patch improves
robustness by converting to the local timezone right before feeding
the value to the datetime-local form input, making it so a different
timezone in OsrdConfState.startTime doesn't break the app.

IsoDateTimeString is not longer used and can be dropped.

Signed-off-by: Simon Ser <contact@emersion.fr>