front: revert tolerances! in StdcmOrigin and StdcmDestination
tolerances are stored in the store and in the local storage. For some users, this field is still undefined in their local storage.
However, since https://github.com/OpenRailAssociation/osrd/pull/9977, we assume that tolerances can not be undefined, which is not true if the
users have not reset their store (=clean their cookies) or manullaly filled it.

To prevent the app from crashing when a user without tolerances opens the stdcm page, I revert the changes in StdcmOrigin and StdcmDestination for now.

Signed-off-by: Clara Ni <clara.ni@outlook.fr>