front: setting power restrictions do not remove PathSteps
When manipulating power restrictions, the pathSteps without any constraints (stop, arrival time or margin) were cleaned.
This was the expected behaviour before, because we could not have pathSteps without constraints.

As the expected behaviour has changed (we now authorize pathSteps without constraints), the pathSteps created from the power restrictions selector
have a new flag and may be cleaned if they are not linked anymore to a power restrictions and have no constraints.
The other pathSteps will not be cleaned.

Signed-off-by: Clara Ni <clara.ni@outlook.fr>