front: use a variable instead of localStorage to restore STDCM form state
Using localStorage here is a bit of a hack: we don't actually want
to persist anything on disk. This also doesn't work when multiple
STDCM tabs are opened: launch 2 simulations, go back to the first
one, click the "Start a new query from the current one" button,
and the state of the second tab gets restored.

Instead, use a global variable to pass the form state to the new
window.

Signed-off-by: Simon Ser <contact@emersion.fr>